requirements.txt
django==3.1.1
djangorestframework>=3.11.1,<3.12.0
psycopg2>=2.8.6,<2.9.0